Updated style.css
Lines 7-13: This sets it up so I can add "var" properties to make the colors appear less redundant.

Due to the root property toward the top of the style, line 8 was changed to be in line 15, where I changed it to point to the right variable.

Lines 14 and 15: changed the property so it points to the right variable. These were moved down to line 21 and 22 respectively.

Line 24 was moved down to line 31, changed the property to the right variable.

Line 31 was moved to line 38, added quotation marks around Calibri

Line 45 was moved to line 52, changed property to variable

Added a comment in style.css for line 85 to show where benefits style begins.

As a result, what was at line 85 is now at line 93, added quotation marks around Calibri

Line 91 was moved to line 99, I changed a property to point to correct variable.

Line 96 was moved to line 104, changed property to point to correct variable.

Line 101 has been moved down to line 109. Changed property to point to correct variable.

Lines 141-143 were moved to 149-151. LIne 149 I added quotation marks around Calibri, lines 150 and 151 I changed properties to point to correct variable.

Lines 150-152 were moved to 158-160.
158: Added quotation marks around Calibri.
159: changed property to point to correct variable.
160: changed property to point to correct variable.

Lines 159-161 were moved to 167-169.
167: added quotation marks around Calibri
168: changed property to point to correct variable
169: changed property to point to correct variable

Added 2 lines at 210 and 211.
210: blank
211: end of footer style comment.