Potentially final commit for challenge.
I believe I have finished this assignment. I added comments so people can see what the var commands do for each instance, did comments about which portion of the style sheet does what in which location. I have tidied up the last of the code, updated the README, added alt tags in the HTML, and the site is unchanged from an end user point of view.